[ refactor ] 변수명, 폴더경로 수정
